refactor: 按照Vue规范重构项目目录结构
