+++ plugins/diff-compare/src/composables/useExcel/index.ts
@@ -1,0 +1,1 @@
+export * from './useExcelDiff'
