+++ plugins/diff-compare/src/composables/useImage/index.ts
@@ -1,0 +1,1 @@
+export * from './useImageDiff'
