+++ plugins/diff-compare/src/composables/usePdf/index.ts
@@ -1,0 +1,1 @@
+export * from './usePdfDiff'
