+++ plugins/diff-compare/src/composables/useText/index.ts
@@ -1,0 +1,3 @@
+export * from './useTextDiff'
+export * from './useSyntaxHighlight'
+export * from './useAutoFormat'
