+++ plugins/diff-compare/src/composables/useText/useAutoFormat.ts
