+++ plugins/diff-compare/src/composables/useText/useSyntaxHighlight.ts
