+++ plugins/diff-compare/src/composables/useWord/index.ts
@@ -1,0 +1,1 @@
+export * from './useWordDiff'
